+++ src/game/visibility/visibility_render.rs
@@ -9,11 +9,11 @@
 // Cas #2c : Range 1 visibility => Quand je me deplace, les cases qui ne sont plus visibles sont bien signalées comme hidden? Oui.
 
 
-use bevy::{prelude::*, utils::HashMap};
+use bevy::prelude::*;
 
 use crate::{engine::render::components::GameMapRender, game::{pieces::components::Npc, tileboard::components::{BoardPosition, Tile}}, vectors::Vector2Int};
 
-use super::components::{ChangeVisibility, ChangeVisibilityStatus, View};
+use super::components::{ChangeVisibility, ChangeVisibilityStatus, RenderVisibilityTile, View};
 
 
  // RENDER_SW corresponds à 0,0.
@@ -27,7 +27,7 @@ const LOGIC_TILE_SE:(i32, i32) = (1, 1);
 const LOGIC_TILE_S:(i32, i32) = (0, 1);
 const LOGIC_TILE_SW:(i32, i32) = (-1, 1);
 const LOGIC_TILE_W:(i32, i32) = (1, 0);
-const LOGIC_TILE_NW:(i32, i32) = (1, -1);
+const LOGIC_TILE_NW:(i32, i32) = (-1, -1);
 const LOGIC_TILE_N:(i32, i32) = (0, -1);
 const LOGIC_TILE_NE:(i32, i32) = (1, -1);
 
@@ -37,13 +37,10 @@ const RENDER_NE_COVER:[(i32,i32); 3] = [LOGIC_TILE_NE, LOGIC_TILE_N, LOGIC_TILE_
 const RENDER_SE_COVER:[(i32,i32); 3] = [LOGIC_TILE_SE, LOGIC_TILE_S, LOGIC_TILE_E];
 
 
-#[derive(Component)]
-pub struct RenderVisibilityTile {
-    pub visibility_score: i32
-}
 
 
 // Cette fonction prends les tuiles logiques dont la visibilité doit changer, et fait les calculs pour determiner quelles tuiles render sont concernées.
+// Une reduction de la duplication de code est possible, cf v2 dans Olds, mais en réalité provoque des ralentissements. P-E une autre facon existe.
 pub fn update_convert_logic_tile_visibility_to_render(
     mut commands: Commands,
     tile_with_change_order_q: Query<(Entity, &ChangeVisibility, &BoardPosition), With<Tile>>,
@@ -192,172 +189,6 @@ pub fn update_tile_visibility_render(
 }
 
 
-
-
-pub fn update_tile_visibility_render_v8a(
-    mut commands: Commands,
-    tile_with_change_order_q: Query<(Entity, &ChangeVisibility, &BoardPosition), With<Tile>>,
-    game_map_render_q: Query<&GameMapRender>, 
-    mut sprite_q: Query<&mut Sprite>,
-    mut visibility_q: Query<&mut Visibility>,
-    view_q: Query<&View>,
- ){
-    let Ok(game_map_render) = game_map_render_q.get_single() else { return; };
-    
-    let mut component_to_delete = Vec::new();
-    let mut entity_change_status = HashMap::new();
-    
-    // Je recupère les entités logiques
-    for (entity, new_visibility, position) in tile_with_change_order_q.iter() {
-        component_to_delete.push(entity);
-          // Je recupere le nouveau statut.
-        let mut visible_status;
-         match new_visibility.new_status {
-            ChangeVisibilityStatus::Visible => visible_status = 1,
-            ChangeVisibilityStatus::Hidden => visible_status = -1,
-        }
-
-        let view = view_q.single();
-        // Si visible, on regarde pour chaque Tuile graphique si les autres tuiles logiques qu'elles couvrent sont visibles.
-        // Sinon, on a le comportement habituel.
-        // SW
-        if visible_status > 0 {
-            let mut score = 0;
-            for logic_tile_around in RENDER_SW_COVER {
-                   if view.visible_tiles.contains(&Vector2Int {x:position.v.x + logic_tile_around.0, y:position.v.y + logic_tile_around.1 } ) {
-                    score += 1;
-                    break;
-                }
-            }
-            if score == 0 {
-                visible_status = -1;    // Elle sera "dans l'obscurité".
-            }
-        }
-        if let Some(render_tile_floor_entity) = game_map_render.floor.get(&Vector2Int { x:position.v.x + RENDER_SW.0, y:position.v.y + RENDER_SW.1 }) {
-            if let Some(render_tile_floor_entity_entry) = entity_change_status.get_mut(render_tile_floor_entity) {
-                *render_tile_floor_entity_entry += visible_status;
-            } else {
-                entity_change_status.insert(render_tile_floor_entity, visible_status);
-            }
-        }
-        if let Some(render_tile_wall_entity) = game_map_render.wall.get(&Vector2Int { x:position.v.x + RENDER_SW.0, y:position.v.y + RENDER_SW.1 }) {
-            if let Some(render_tile_wall_entity_entry) = entity_change_status.get_mut(render_tile_wall_entity) {
-                *render_tile_wall_entity_entry += visible_status;
-            } else {
-                entity_change_status.insert(render_tile_wall_entity, visible_status);                
-            }
-        }
-
-        //RENDER_NW
-        if visible_status > 0 {
-            let mut score = 0;
-            for logic_tile_around in RENDER_NW_COVER {
-                   if view.visible_tiles.contains(&Vector2Int {x:position.v.x + logic_tile_around.0, y:position.v.y + logic_tile_around.1 } ) {
-                    score += 1;
-                    break;
-                }
-            }
-            if score == 0 {
-                visible_status = -1;    // Elle sera "dans l'obscurité".
-            }
-        }
-        if let Some(render_tile_floor_entity) = game_map_render.floor.get(&Vector2Int { x:position.v.x + RENDER_NW.0, y:position.v.y + RENDER_NW.1 }) {
-            if let Some(render_tile_floor_entity_entry) = entity_change_status.get_mut(render_tile_floor_entity) {
-                *render_tile_floor_entity_entry += visible_status;
-            } else {
-                entity_change_status.insert(render_tile_floor_entity, visible_status);
-            }
-        }
-        if let Some(render_tile_wall_entity) = game_map_render.wall.get(&Vector2Int { x:position.v.x + RENDER_NW.0, y:position.v.y + RENDER_NW.1 }) {
-            if let Some(render_tile_wall_entity_entry) = entity_change_status.get_mut(render_tile_wall_entity) {
-                *render_tile_wall_entity_entry += visible_status;
-            } else {
-                entity_change_status.insert(render_tile_wall_entity, visible_status);                
-            }
-        }
-
-        //RENDER_NE        
-        if visible_status > 0 {
-            let mut score = 0;
-            for logic_tile_around in RENDER_NE_COVER {
-                   if view.visible_tiles.contains(&Vector2Int {x:position.v.x + logic_tile_around.0, y:position.v.y + logic_tile_around.1 } ) {
-                    score += 1;
-                    break;
-                }
-            }
-            if score == 0 {
-                visible_status = -1;    // Elle sera "dans l'obscurité".
-            }
-        }
-        if let Some(render_tile_floor_entity) = game_map_render.floor.get(&Vector2Int { x:position.v.x + RENDER_NE.0, y:position.v.y + RENDER_NE.1 }) {
-            if let Some(render_tile_floor_entity_entry) = entity_change_status.get_mut(render_tile_floor_entity) {
-                *render_tile_floor_entity_entry += visible_status;
-            } else {
-                entity_change_status.insert(render_tile_floor_entity, visible_status);
-            }
-        }
-        if let Some(render_tile_wall_entity) = game_map_render.wall.get(&Vector2Int { x:position.v.x + RENDER_NE.0, y:position.v.y + RENDER_NE.1 }) {
-            if let Some(render_tile_wall_entity_entry) = entity_change_status.get_mut(render_tile_wall_entity) {
-                *render_tile_wall_entity_entry += visible_status;
-            } else {
-                entity_change_status.insert(render_tile_wall_entity, visible_status);                
-            }
-        }
-
-        //RENDER_SE
-        if visible_status > 0 {
-            let mut score = 0;
-            for logic_tile_around in RENDER_SE_COVER {
-                   if view.visible_tiles.contains(&Vector2Int {x:position.v.x + logic_tile_around.0, y:position.v.y + logic_tile_around.1 } ) {
-                    score += 1;
-                    break;
-                }
-            }
-            if score == 0 {
-                visible_status = -1;    // Elle sera "dans l'obscurité".
-            }
-        }
-        if let Some(render_tile_floor_entity) = game_map_render.floor.get(&Vector2Int { x:position.v.x + RENDER_SE.0, y:position.v.y + RENDER_SE.1 }) {
-            if let Some(render_tile_floor_entity_entry) = entity_change_status.get_mut(render_tile_floor_entity) {
-                *render_tile_floor_entity_entry += visible_status;
-            } else {
-                entity_change_status.insert(render_tile_floor_entity, visible_status);
-            }
-        }
-        if let Some(render_tile_wall_entity) = game_map_render.wall.get(&Vector2Int { x:position.v.x + RENDER_SE.0, y:position.v.y + RENDER_SE.1 }) {
-            if let Some(render_tile_wall_entity_entry) = entity_change_status.get_mut(render_tile_wall_entity) {
-                *render_tile_wall_entity_entry += visible_status;
-            } else {
-                entity_change_status.insert(render_tile_wall_entity, visible_status);                
-            }
-        }
-    }
-    for (entity, score) in entity_change_status {
-        if let Ok(mut visibility) = visibility_q.get_mut(*entity) {
-            * visibility = Visibility::Visible;
-            // Hidden but known
-            if score < 0 {
-                if let Ok(mut sprite) = sprite_q.get_mut(*entity){
-                    sprite.color.set_a(0.5);
-                    sprite.color.set_r(0.5);
-                    sprite.color.set_g(0.5);
-                    sprite.color.set_b(0.5);
-                }
-            // visible
-            } else {
-                if let Ok(mut sprite) = sprite_q.get_mut(*entity){
-                    sprite.color.set_a(1.0);
-                    sprite.color.set_r(1.0);
-                    sprite.color.set_g(1.0);
-                    sprite.color.set_b(1.0);
-                }
-            }
-        }
-    }
-    for entity in component_to_delete {
-        commands.entity(entity).remove::<ChangeVisibility>();
-    }
- }
